Restructure for shared runtime in rsc. Also unify code paths for getTranslations -> getTranslator
