+++ packages/next-intl/src/navigation/react-server/redirects.tsx
@@ -1,4 +1,4 @@
-import {getRequestLocale} from '../../server/react-server/RequestLocale';
+import {getRequestLocale} from '../../runtimes/react-server.shared-runtime';
 import {ParametersExceptFirst} from '../../shared/types';
 import {baseRedirect, basePermanentRedirect} from '../shared/redirects';
