+++ packages/next-intl/src/react-server/NextIntlClientProviderServer.tsx
@@ -1,5 +1,9 @@
 import React, {ComponentProps} from 'react';
-import {getLocale, getNow, getTimeZone} from '../server.react-server';
+import {
+  getLocale,
+  getNow,
+  getTimeZone
+} from '../runtimes/react-server.shared-runtime';
 import BaseNextIntlClientProvider from '../shared/NextIntlClientProvider';
 
 type Props = ComponentProps<typeof BaseNextIntlClientProvider>;
