+++ packages/next-intl/src/react-server/useConfig.tsx
@@ -1,5 +1,5 @@
 import {use} from 'react';
-import getConfig from '../server/react-server/getConfig';
+import {getConfig} from '../runtimes/react-server.shared-runtime';
 
 function useHook<Value>(hookName: string, promise: Promise<Value>) {
   try {
