+++ packages/next-intl/src/react-server/useMessages.tsx
@@ -1,5 +1,5 @@
 import type {useMessages as useMessagesType} from 'use-intl';
-import {getMessagesFromConfig} from '../server/react-server/getMessages';
+import {getMessagesFromConfig} from '../runtimes/react-server';
 import useConfig from './useConfig';
 
 export default function useMessages(
