+++ packages/next-intl/src/react-server/useTranslations.tsx
@@ -1,10 +1,10 @@
 import type {useTranslations as useTranslationsType} from 'use-intl';
-import getBaseTranslator from './getTranslator';
+import {getTranslator} from '../runtimes/react-server';
 import useConfig from './useConfig';
 
 export default function useTranslations(
   ...[namespace]: Parameters<typeof useTranslationsType>
 ): ReturnType<typeof useTranslationsType> {
   const config = useConfig('useTranslations');
-  return getBaseTranslator(config, namespace);
+  return getTranslator(config, namespace);
 }
