+++ packages/next-intl/src/runtimes/react-server.shared-runtime.tsx
@@ -1,0 +1,3 @@
+// This entry point needs to be shared to enable caching and code sharing
+// across `next-intl/server` (react-server) and `next-intl` (react-server)
+export * from './react-server';
