+++ packages/next-intl/src/runtimes/react-server/getConfig.tsx
