+++ packages/next-intl/src/runtimes/react-server/getMessageFormatCache.tsx
@@ -1,0 +1,8 @@
+import {cache} from 'react';
+
+function getMessageFormatCacheImpl() {
+  return new Map();
+}
+const getMessageFormatCache = cache(getMessageFormatCacheImpl);
+
+export default getMessageFormatCache;
