+++ packages/next-intl/src/runtimes/react-server/getTranslations.tsx
@@ -1,6 +1,5 @@
 import {ReactElement, ReactNodeArray, cache} from 'react';
 import {
-  createTranslator,
   Formats,
   TranslationValues,
   MessageKeys,
@@ -10,8 +9,8 @@ import {
   RichTranslationValues,
   MarkupTranslationValues
 } from 'use-intl/core';
-import {getMessageFormatCache} from '../../shared/messageFormatCache';
 import getConfig from './getConfig';
+import getTranslator from './getTranslator';
 
 // Maintainer note: `getTranslations` has two different call signatures.
 // We need to define these with function overloads, otherwise TypeScript
@@ -212,13 +211,7 @@ async function getTranslations<
   }
 
   const config = await getConfig(locale);
-
-  return createTranslator({
-    ...config,
-    messageFormatCache: getMessageFormatCache(),
-    namespace,
-    messages: config.messages
-  });
+  return getTranslator(config, namespace);
 }
 
 export default cache(getTranslations);
