+++ packages/next-intl/src/server/react-server/index.tsx
@@ -2,12 +2,13 @@
  * Server-only APIs available via `next-intl/server`.
  */
 
-export {default as getRequestConfig} from './getRequestConfig';
-export {default as getFormatter} from './getFormatter';
-export {default as getNow} from './getNow';
-export {default as getTimeZone} from './getTimeZone';
-export {default as getTranslations} from './getTranslations';
-export {default as getMessages} from './getMessages';
-export {default as getLocale} from './getLocale';
-
-export {setRequestLocale as unstable_setRequestLocale} from './RequestLocale';
+export {
+  setRequestLocale as unstable_setRequestLocale,
+  getRequestConfig,
+  getFormatter,
+  getNow,
+  getTimeZone,
+  getTranslations,
+  getMessages,
+  getLocale
+} from '../../runtimes/react-server.shared-runtime';
