--- packages/next-intl/src/shared/messageFormatCache.tsx
@@ -1,6 +1,0 @@
-import {cache} from 'react';
-
-function getMessageFormatCacheImpl() {
-  return new Map();
-}
-export const getMessageFormatCache = cache(getMessageFormatCacheImpl);
